audio: fix audio playback with small period sizes
BZ: 3559
split common firmware project
keep history as much as possible

For audio playback with odd period sizes silence was getting inserted for the last 1 sec
playback before the dma latencybuffer worth of data was played. So on that 1 sec when
vol control was applied, it was giving glitch noise. TO fix this problem dma latency buffer
size was changed to 4ms from 2ms.

Change-Id: I9c41267af79df7b6dec88ef57046b06c08f57ab6
Signed-off-by: Rohit Kumar Sindhu <rohit.k.sindhu@intel.com>
Signed-off-by: Erwan Mahe <erwanx.mahe@intel.com>