[REBASED on main R3 from r2-stable branch]audio_fw: LPE FW v3.0.24 provides fix for pb noise and mem corruption issues
BZ: 17975

Bug 15936 - [Audio] Streaming sound becomes abnormal after recording with
Sound Recorder. This was caused due to un-initialization of variable which
retains old value resulting noisy playback.The problem was found during
klocwork analysis of the LPE code.

Bug 16834 - crash when applying less than 32 taps in equalizer FIR filter.
There was a bug in Mono and Stereo EQ modules, that when the number of taps
for FIR filter was changed, FIR delay field was not updated in the set_param()
function. This was causing memory corruption inside the MONO EQ assembly code.

Change-Id: I77991d02af1415a4fcdfbb61cce3e8b4486bd6ab
Old-Change-Id: I114032cff881edb0d024f70aaf779e36547daf0b
Signed-off-by: Sanyog <sanyog.r.kale@intel.com>
Reviewed-on: http://android.intel.com:8080/29710
Reviewed-by: M, Arulselvan <arulselvan.m@intel.com>
Tested-by: M, Arulselvan <arulselvan.m@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>