[Rebase-R2-R3]audio_fw: LPE FW v3.0.25 provides fix for no audio in some scenarios of audio playback
BZ: 19267

Old-ID R2: Bug 18108 - [IMAGING] Shutter sound not heard when taking a photo.
The issue was with WAITI implementation in firmware due to which in some
cases firmware stuck in WAITI and delays coming out of WAITI resulting into
no audio. Currently WAITI is disabled in firmware and issue is not observed.

Change-Id: Icb32d4bd17048cc8f8a716333516d88c771be998
Old-Change-Id: I1f3dc99568d049756418722d3fff2d98e960a1d9
Signed-off-by: Jeeja Kp <jeeja.kp@intel.com>
Reviewed-on: http://android.intel.com:8080/31046
Tested-by: M, Arulselvan <arulselvan.m@intel.com>
Reviewed-by: M, Arulselvan <arulselvan.m@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>