audio_fw: LPE FW v5.02.01 Providing Fix for the Write-error during the simultaneous playback and capture.
BZ: 34409

Bug 34409 - Sound recording works first 2 sec and jump to 31 sec.

In the FW, the mixer mask is cleared during the Drop command from IA.
Driver does not provide the mixer controls across Drop and Start
commands. This results in the WriteError on alsa and resulted in this
issue. The delay of 30 Seconds is because of the slow recovery on
Alsa.

This FW patch will fix the write error from Alsa and henceforth
never triggers the alsa recovery.

Change-Id: Ife5abd97a5c6fd484ea52d7fd175aa2c66f34bb3
Signed-off-by: Sanyog <sanyog.r.kale@intel.com>