[PORT FROM R3]audio_fw:LPE FW v03.00.35 add dual mono support for ihf streams
BZ: 38078

Bug 38078 - Dual Mono implementation for IHF streams when IHF
stream  is configured as Mono.

If IHF pipeline is configured to Mono, renderer module copies
the content of Left channel on to Right channel to make it as a dual mono.

In corner scenarios delay field could become negative. This firmware
fixes this issue by validating delay field before updating it.

Change-Id: Id55829e688c4f441c8227c999d8bb288e7dc6fca
Signed-off-by: Dinesh Mirche <dinesh.mirche@intel.com>