audio_fw: LPE FW v5.06.05 Fix to resolve no volume in right channel in offload path post mute/unmute.
BZ: 81580

[CTP-offload]Sound lost in right channel of wsHS when HDMI plug in
during playing mp3 file with wsHS.

There was no audio heard in the right channel when the volume was ramped
from mute. This was because there is a flag to indicate whether the target
gain has been reached by the channel, this flag was set only for left
channel and not for right. Changing this in volume library solved the issue.

Change-Id: Ida33ac36e30bd8303aa2e0ee8048cfb8951e1bf1
Signed-off-by: Pankaj Mistry <pankaj.mistry@intel.com>