audio_fw: LPE FW v05.06.14 Fix for 32bit Left justified support in firmware
BZ:101860

Bug 101860 : Noise can be heard regularly when
playback every 38.8 seconds

In this FW,changes are made to support 32 bit
left justified format for rendering and
capturing data to and from codec.

This FW has changes to support 32 bit left
justified format for rendering and capturing
data to and from codec.

Change-Id: I371ad9b97f707c6e94bf66d36fd800e92831a79c
Signed-off-by: vijayhs <vijay.h.s@intel.com>