audio_fw:LPE FW v05.07.02 Fix for ANR during stop media and start media operations
BZ:115846

Bug 115846: ANR when media server died during GMS
music playback

In this FW,changes are made to save the mixer settings
when the LPE enters D3 and restore mixer settings when
LPE comes out of D3.

This FW has changes to save the mixer settings on LPE
entering D3 and restore the mixer settings on coming out
of D3.

Change-Id: I9a3e79794beaf7d76a94562e46a6ccb1aea70cc4
Signed-off-by: vijayhs <vijay.h.s@intel.com>