audio_fw:LPE FW v05.07.07 Fixed fabric error due to DMA response timeout issue
BZ: 130643

Bug 130643 - [TW 824] [Customer platform only] Audio: Watchdog reset.

When multiple streams are active, we switch the LPE clock from 19.2
to 200Mhz based on the number of streams present. During this switch
we suspend any ongoing DMA transactions and send LPE to waiti state so
that clock switching can go through(Clocks cannot be switched when LPE
is running).After setting the suspend bit,we find that the current DMA
never completes and we are struck in a while loop and there by a fabric occurs.

Instead of suspending the DMA channels midway through the transaction, wait
till current DMA is complete and then switch the clock which causes an additional
delay of 50 micro seconds as compared to direct suspend.

md5sum : 18ce911d9423716a4d01dacb182a2179

Change-Id: I5e881c89989be78fd448540234ac9dc85567c746
Signed-off-by: mirche dinesh <dinesh.mirche@intel.com>