audio_fw:LPE FW v8.0A.02 Fixes low volume issue.
BZ: 160218

Bug 160218 - [FFRD8] Video and other apps have audio too low[IHF Volume is very Low]

This Firmware fixes the low IHF volume issue. The issue was because
restoring the context was not proper in the firmware. Firmware was saving to and
restoring the context from DDR data section. When driver Re downloads the firmware it
resets the DDR data section as well. So when firmware restores the context, data was lost.

md5sum for fw binary : 2b212d36bc115f730ddd6163d392c740

Change-Id: I60aa62f1c6a5b7fb327fe6cc396140e3b57593e8
Signed-off-by: mirche dinesh <dinesh.mirche@intel.com>