[PORT FROM R44B-STABLE]audio_fw:LPE FW v8.0A.04 Fixes IHF low volume and sudden increase in volume
BZ: 160218 169124

Bug 160218 - [FFRD8] Video and other apps have audio too low
[IHF Volume is very Low]
Bug 169124 - The volume of video/audio increase suddenly when change the
volume level to 1 in WIDI BG mode.

This Firmware fixes the low IHF volume issue. The issue was because
restoring the context was not proper in the firmware. Firmware was saving
to and restoring the context from DDR data section. When driver Re downloads
the firmware it resets the DDR data section as well. So when firmware
restores the context, data was lost. To fix this issue context is saved in
predefined memory in DDR.

This Firmware has a fix in volume library. This volume library ensures
that correct ramp steps are seteven when current gain is same as final gain.
The sudden increase in volume issue is fixed only for offload path.

md5sum for binary : 67f88e5ae3884b10216ace67e7f5681c

Change-Id: I6cad6df58f6e34b374d60501e9ba80c033d3d3b0
Orig-Change-Id: Ib48a2ac716d000dc7d3c43bb2edd02309e30db59
Signed-off-by: mirche dinesh <dinesh.mirche@intel.com>